refactor: Move response.rs tests to separate file
Per CLAUDE.md guideline: "Put unit tests in separate files rather than
inline #[cfg(test)] mod tests blocks."

This refactor moves ~280 lines of test code from
src/bin/midtown/cli/response.rs to a new response_tests.rs file,
following the pattern used in the daemon module.

Changes:
- Created src/bin/midtown/cli/response_tests.rs with all test functions
- Replaced inline test module with #[path = "response_tests.rs"] directive
- All tests pass unchanged

Note: WebSocket CoworkerStatusData changes deferred to task !1226 per
lead's direction.
